refactor: 💡 move exceptions in new packages
